Add remove button on book admin page
